refactor(gpu_gkr_model): split storage_layout into by-concern submodules
Pure relocation: mod.rs's monolithic ~700 lines are moved verbatim into
types.rs (data model), construct.rs (build from an artifact), alias.rs
(CopyIn*Field alias resolution), tower.rs (dim-reducing tower layers), and
fixtures.rs (hand-crafted test-support layout). mod.rs becomes a thin facade
re-exporting the same storage_layout::{FieldType, StorageSlot,
GpuGKRLayerLayout, GpuGKRStorageLayout, address_storage_layer,
handcrafted_layout} paths gpu_circuit_prover's gkr facade depends on.
No logic changes; only the minimal pub(super) visibility bumps and import
adjustments the file split requires.
